Code re-factoring. SpadePanel class moved out of Cytospade.java to a new file SpadePanel.java. FCScomputations.java created to handle computations related to data extrated from FCS file.
